Move class to separate file
